requirements_predict.txt
torch
torchvision
neptune-client
python-dotenv
Pillow
argparse
